feat(output): allow HTML-only Unicode shipout
Propagate DVI output selection through execution, incremental sessions, native drivers, and WASM so HTML-only OpenType pages retain full Unicode without constructing TeX82 plans. Keep DVI byte capability failures unchanged when DVI is requested.\n\nWiden positioned text units and DOM source mapping to Unicode scalars, preserve separate physical byte codes for classic outputs, and cover native, WASM, JavaScript, Firefox, and Chromium paths.

+++ crates/tex-out/src/html/tests.rs
@@ -446,7 +446,7 @@ fn unavailable_text_mapping_is_actionable() {
         error,
         HtmlError::MissingTextMapping {
             font: "cmr10".to_owned(),
-            code: b'B'
+            code: u32::from(b'B')
         }
     );
 }
